fix: moved all icons to pure svgs
import like any other typ. vue component
