+++ package.json
@@ -45,6 +45,7 @@
     "vite": "^4.4.7",
     "vite-plugin-dts": "^3.3.1",
     "vite-plugin-vue-markdown": "^0.23.7",
+    "vite-svg-loader": "^4.0.0",
     "vue-tsc": "^1.8.8"
   },
   "peerDependencies": {
