+++ src/components/account/account-unlink.vue
@@ -4,11 +4,11 @@
   -->
 
 <script setup lang="ts">
-import CrossMarkIconOutline from "@/components/icons/cross-mark-icon-outline.vue";
+import CrossMarkIconOutline from "@/assets/icons/cross-mark-outline.svg";
 import CircleButton from "@/components/buttons/circle-button.vue";
 import HeaderBack from "@/components/header/header-back.vue";
 import AccountForm from "@/components/account/account-form.vue";
-import AccountIconOutline from "@/components/icons/account-icon-outline.vue";
+import AccountIconOutline from "@/assets/icons/account-outline.svg";
 import TextButton from "@/components/buttons/text-button.vue";
 import { inject, ref } from "vue";
 import {
