--- src/components/icons/account-icon-outline.vue
@@ -1,16 +1,0 @@
-<!--
-  - Copyright (c) TIKI Inc.
-  - MIT license. See LICENSE file in root directory.
-  -->
-
-<template>
-  <svg
-    viewBox="0 0 100 100"
-    fill="currentColor"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M0 50C0 22.3858 22.3858 0 50 0C77.6142 0 100 22.3857 100 50C100 64.8086 93.5585 78.1165 83.3341 87.2676C74.4902 95.1834 62.8039 100 50 100C37.1961 100 25.5099 95.1834 16.6659 87.2676C6.44148 78.1165 0 64.8086 0 50ZM50 5.88235C25.6345 5.88235 5.88235 25.6345 5.88235 50C5.88235 61.0261 9.92459 71.1052 16.6137 78.8408C17.2375 76.8927 18.1612 75.1359 19.3389 73.5669C21.6295 70.5152 24.7681 68.3208 28.182 66.7494C34.9519 63.6331 43.2874 62.7451 50 62.7451C56.7126 62.7451 65.0481 63.6331 71.818 66.7494C75.2319 68.3208 78.3705 70.5152 80.6611 73.5669C81.8388 75.1359 82.7625 76.8927 83.3863 78.8408C90.0754 71.1052 94.1176 61.0261 94.1176 50C94.1176 25.6345 74.3654 5.88235 50 5.88235ZM78.3839 83.7759C78.1738 81.0324 77.2799 78.8612 75.9565 77.0981C74.4072 75.0339 72.1536 73.3795 69.3584 72.0928C63.7099 69.4928 56.3592 68.6275 50 68.6275C43.6408 68.6275 36.2901 69.4928 30.6416 72.0928C27.8464 73.3795 25.5928 75.0339 24.0435 77.0981C22.7201 78.8612 21.8262 81.0324 21.6161 83.7759C29.2895 90.2317 39.1878 94.1176 50 94.1176C60.8122 94.1176 70.7106 90.2317 78.3839 83.7759ZM50 21.5686C42.9612 21.5686 37.2549 27.2748 37.2549 34.3137C37.2549 41.3525 42.9612 47.0588 50 47.0588C57.0388 47.0588 62.7451 41.3525 62.7451 34.3137C62.7451 27.2748 57.0388 21.5686 50 21.5686ZM31.3725 34.3137C31.3725 24.026 39.7125 15.6863 50 15.6863C60.2875 15.6863 68.6275 24.026 68.6275 34.3137C68.6275 44.6012 60.2875 52.9412 50 52.9412C39.7125 52.9412 31.3725 44.6012 31.3725 34.3137Z"
-    />
-  </svg>
-</template>
