--- src/components/icons/account-icon.vue
@@ -1,19 +1,0 @@
-<!--
-  - Copyright (c) TIKI Inc.
-  - MIT license. See LICENSE file in root directory.
-  -->
-
-<template>
-  <svg
-    viewBox="0 0 100 100"
-    fill="currentColor"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M32.3625 33.8188C32.3625 24.0779 40.259 16.1812 50 16.1812C59.7409 16.1812 67.6375 24.0779 67.6375 33.8188C67.6375 43.5598 59.7409 51.4563 50 51.4563C40.259 51.4563 32.3625 43.5598 32.3625 33.8188Z"
-    />
-    <path
-      d="M0 50C0 22.3858 22.3858 0 50 0C77.6142 0 100 22.3858 100 50C100 64.8085 93.5606 78.1148 83.3338 87.2682C74.4887 95.1847 62.805 100 50 100C37.195 100 25.5113 95.1847 16.6663 87.2682C6.43948 78.1148 0 64.8085 0 50ZM50 2.91262C23.9944 2.91262 2.91262 23.9944 2.91262 50C2.91262 62.8548 8.06214 74.505 16.4145 83.0035C16.875 80.0033 18 77.4334 19.6332 75.2575C21.7973 72.3744 24.7975 70.2555 28.1532 68.7109C34.8359 65.6348 43.1703 64.7249 50 64.7249C56.8297 64.7249 65.1641 65.6348 71.8468 68.7109C75.2025 70.2555 78.2027 72.3744 80.3668 75.2575C82 77.4334 83.1251 80.0033 83.5854 83.0037C91.9379 74.505 97.0874 62.8548 97.0874 50C97.0874 23.9944 76.0056 2.91262 50 2.91262Z"
-    />
-  </svg>
-</template>
