--- src/components/icons/arrow-icon.vue
@@ -1,16 +1,0 @@
-<!--
-  - Copyright (c) TIKI Inc.
-  - MIT license. See LICENSE file in root directory.
-  -->
-
-<template>
-  <svg
-    viewBox="0 0 100 100"
-    fill="currentColor"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M50.0009 100L58.8353 91.166L23.953 56.1363H100.001V43.864H23.953L58.8353 8.83448L50.0009 0L0 50.0009L50.0009 100Z"
-    />
-  </svg>
-</template>
