--- src/components/icons/check-icon.vue
@@ -1,16 +1,0 @@
-<!--
-  - Copyright (c) TIKI Inc.
-  - MIT license. See LICENSE file in root directory.
-  -->
-
-<template>
-  <svg
-    viewBox="0 0 100 100"
-    fill="currentColor"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M95.633 6.67697C91.1205 4.24589 86.0874 8.9344 83.1369 11.7128C76.3684 18.3114 70.641 25.952 64.2191 32.8979C57.1032 40.5385 50.508 48.179 43.2186 55.6462C39.0532 59.8137 34.5407 64.3286 31.7638 69.5381C25.5157 63.4601 20.1354 56.8614 13.1931 51.4786C8.15989 47.6583 -0.170889 44.88 0.00266917 54.0833C0.349785 66.0654 10.9368 78.9154 18.7469 87.0766C22.0445 90.5495 26.3835 94.1962 31.4166 94.3698C37.4912 94.7171 43.7392 87.4239 47.384 83.4299C53.8059 76.484 59.0126 68.6695 64.9133 61.5502C72.5498 52.1732 80.36 42.9695 87.8229 33.4189C92.509 27.5148 107.261 12.928 95.633 6.67697Z"
-    />
-  </svg>
-</template>
