--- src/components/icons/cross-mark-icon-outline.vue
@@ -1,16 +1,0 @@
-<!--
-  - Copyright (c) TIKI Inc.
-  - MIT license. See LICENSE file in root directory.
-  -->
-
-<template>
-  <svg
-    viewBox="0 0 100 100"
-    fill="currentColor"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M3.41274 96.5871C5.00663 98.1821 7.09403 98.9793 9.18315 98.9793C11.2723 98.9793 13.3615 98.1821 14.9554 96.5871L49.9994 61.5427L85.0443 96.5879C86.6386 98.1836 88.7279 98.9807 90.8171 98.9807C92.9057 98.9807 94.9936 98.1836 96.5871 96.5879C99.7771 93.4 99.7771 88.235 96.5871 85.045L61.5421 50.0001L96.5871 14.9549C99.7764 11.7671 99.775 6.60008 96.5871 3.4123C93.3993 0.222776 88.2343 0.222776 85.0443 3.4123L49.9994 38.4575L14.9554 3.41339C11.7676 0.223867 6.60052 0.225611 3.41274 3.41339C0.223219 6.60117 0.223219 11.7665 3.41274 14.956L38.4569 50.0001L3.41274 85.0443C0.223219 88.2321 0.223219 93.3993 3.41274 96.5871Z"
-    />
-  </svg>
-</template>
