--- src/components/icons/cross-mark-icon.vue
@@ -1,16 +1,0 @@
-<!--
-  - Copyright (c) TIKI Inc.
-  - MIT license. See LICENSE file in root directory.
-  -->
-
-<template>
-  <svg
-    viewBox="0 0 100 100"
-    fill="currentColor"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M94.1058 73.2548C91.5492 70.4035 88.7875 67.7222 86.0917 65.0107C84.0842 63.0033 81.8718 61.1757 80.0192 59.0534C75.0708 53.3159 70.2521 47.4337 65.2687 41.4765C65.3286 41.3167 65.3935 41.1619 65.4734 41.0171C71.3507 33.4271 77.6824 26.1962 84.4333 19.3751C85.4425 18.3764 86.5158 17.4327 87.4642 16.399C89.5867 14.0621 89.7167 12.4991 87.5942 10.2871C85.9117 8.4595 83.9942 6.84662 81.9019 5.50335C77.7623 2.94671 73.0935 1.66838 68.5051 0.110363C67.7011 -0.194236 66.8023 0.150311 66.4078 0.914317C64.5902 3.38109 62.6727 5.77296 60.7453 8.14987C56.9103 12.8787 53.0753 17.5725 49.2403 22.2868C49.0007 22.6064 48.7759 22.941 48.5762 23.2855C42.5541 16.7792 37.1861 10.2326 31.1642 4.22574C30.1105 4.74507 29.1118 5.37922 28.198 6.1183C26.2057 8.24057 24.3631 10.4626 22.4605 12.6996C21.0174 14.4274 19.6492 16.215 18.241 17.9677C16.8328 19.7204 15.5445 21.1486 14.2912 22.8013C12.8982 24.6538 13.0129 25.7275 14.6857 27.3902C18.3709 31.2502 22.3507 34.8255 26.5902 38.0762C28.9271 39.7491 30.9893 41.8114 33.2015 43.7388C27.2143 52.8768 21.6017 62.0048 14.7259 70.2142C11.5052 74.0492 8.53908 78.0889 5.50801 82.0636C4.22968 83.7917 4.22968 84.9 5.80263 86.4875C6.61656 87.3217 7.51533 88.0758 8.47408 88.735C12.7834 91.6008 16.7982 94.8867 20.4633 98.5417C21.1474 99.1058 21.8865 99.5958 22.6755 100C23.2498 98.7067 23.9138 97.4633 24.6579 96.265C32.8222 84.845 41.0167 73.4647 49.2401 62.1196C49.5847 61.6452 49.9542 61.2257 50.3786 60.7114C50.813 61.0959 51.1826 61.4154 51.5421 61.76L63.8161 74.2486C67.2266 77.699 70.6321 81.1445 74.0427 84.575C76.0251 86.5825 76.8939 86.5825 79.2858 85.2792C83.77 82.8023 88.3292 80.4204 92.9575 78.1385C95.7342 76.7553 96.2433 75.5419 94.1058 73.2548Z"
-    />
-  </svg>
-</template>
