--- src/components/icons/question-icon.vue
@@ -1,19 +1,0 @@
-<!--
-  - Copyright (c) TIKI Inc.
-  - MIT license. See LICENSE file in root directory.
-  -->
-
-<template>
-  <svg
-    viewBox="0 0 100 100"
-    fill="currentColor"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M49.5004 76.0794C45.4884 76.0794 42.236 72.8424 42.236 68.8406C43.18 59.6371 47.8926 51.2281 55.2602 45.6075C59.7662 41.1125 64.0216 36.8677 64.0216 32.6228C64.0216 24.6186 57.5242 18.1373 49.5002 18.1373C41.4762 18.1373 34.9789 24.6186 34.9789 32.6228C34.9789 36.6249 31.7264 39.862 27.7144 39.862C23.7024 39.862 20.4573 36.6249 20.4573 32.6228C20.4573 22.2718 25.9886 12.7009 34.9787 7.52923C43.9614 2.35008 55.0387 2.35008 64.0218 7.52923C73.0119 12.7011 78.5432 22.2722 78.5432 32.6228C77.6065 41.8262 72.8939 50.2277 65.5263 55.8482C61.0203 60.3506 56.7649 64.5953 56.7649 68.84C56.7649 70.76 55.9979 72.5994 54.6335 73.9606C53.2765 75.3218 51.4255 76.0794 49.5004 76.0794Z"
-    />
-    <path
-      d="M56.7729 90.5753C56.7729 100.239 42.2366 100.239 42.2366 90.5753C42.2366 80.9106 56.7729 80.9106 56.7729 90.5753Z"
-    />
-  </svg>
-</template>
