--- src/components/icons/receipt-icon.vue
@@ -1,16 +1,0 @@
-<!--
-  - Copyright (c) TIKI Inc.
-  - MIT license. See LICENSE file in root directory.
-  -->
-
-<template>
-  <svg
-    viewBox="0 0 100 100"
-    fill="currentColor"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M82.9661 0H13.9297C12.3117 0 11 1.31172 11 2.92968V88.4406C11 89.2172 11.3088 89.9628 11.858 90.5122L20.4877 99.1417C21.6318 100.286 23.4867 100.286 24.6311 99.1417L31.189 92.5839L37.7466 99.1417C38.2961 99.6911 39.0412 100 39.8181 100C40.5951 100 41.3404 99.6917 41.8898 99.1417L48.4476 92.5839L55.0056 99.1417C55.555 99.6911 56.3003 100 57.0771 100C57.854 100 58.5993 99.6917 59.1487 99.1417L65.7066 92.5839L72.2644 99.1417C72.8367 99.7139 73.5861 100 74.3361 100C75.0856 100 75.8356 99.7139 76.4072 99.1417L85.0372 90.5122C85.5867 89.9628 85.8956 89.2178 85.8956 88.4406V2.92968C85.8961 1.31172 84.5839 0 82.9661 0ZM45.5177 22.006V18.6199C45.5177 17.0019 46.8296 15.6902 48.4474 15.6902C50.0656 15.6902 51.3771 17.0019 51.3771 18.6199V21.9662H54.7237C56.3415 21.9662 57.6534 23.2779 57.6534 24.8958C57.6534 26.5138 56.3415 27.8256 54.7237 27.8256H46.3558C44.5107 27.8256 43.0095 29.3267 43.0095 31.1718C43.0095 33.0171 44.5107 34.5185 46.3558 34.5185H50.5399C55.6159 34.5185 59.7458 38.6479 59.7458 43.7241C59.7458 48.5177 56.0622 52.4655 51.3771 52.8901V56.2761C51.3771 57.8944 50.0656 59.2061 48.4474 59.2061C46.8296 59.2061 45.5177 57.8944 45.5177 56.2761V52.9296H42.1714C40.5537 52.9296 39.2417 51.6179 39.2417 49.9999C39.2417 48.3819 40.5537 47.0702 42.1714 47.0702H50.5399C52.3851 47.0702 53.8864 45.5691 53.8864 43.7239C53.8864 41.8788 52.3853 40.3777 50.5399 40.3777H46.3558C41.2796 40.3777 37.1502 36.2478 37.1502 31.1716C37.1503 26.3783 40.8333 22.4308 45.5177 22.006ZM64.138 74.8956H32.7578C31.1398 74.8956 29.8281 73.5839 29.8281 71.9661C29.8281 70.3483 31.1398 69.0367 32.7578 69.0367H64.138C65.7558 69.0367 67.0678 70.3483 67.0678 71.9661C67.0678 73.5839 65.7558 74.8956 64.138 74.8956Z"
-    />
-  </svg>
-</template>
