--- src/components/icons/redeem-icon.vue
@@ -1,19 +1,0 @@
-<!--
-  - Copyright (c) TIKI Inc.
-  - MIT license. See LICENSE file in root directory.
-  -->
-
-<template>
-  <svg
-    viewBox="0 0 100 100"
-    fill="currentColor"
-    stroke="currentColor"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      fill-rule="evenodd"
-      d="M28.0782 29.2738C23.5339 26.5754 20.6664 21.5194 20.6372 16.1926C20.6588 11.064 23.2666 6.25989 27.5512 3.4467C31.3273 0.968975 36.1961 0.330601 40.4915 1.72637C44.7869 3.12218 48.2819 6.37894 49.9914 10.5662L49.9931 10.5707C50.2064 10.0368 50.4506 9.51481 50.7258 9.00794C52.8356 5.10203 56.7202 2.20234 61.0695 1.32592C65.5814 0.41702 70.2447 1.56393 73.8153 4.46358C77.3209 7.29836 79.3441 11.6911 79.3659 16.1922C79.3441 19.5572 78.2404 22.8358 76.174 25.4975C74.9974 27.0193 73.5452 28.298 71.9047 29.2738H92.0693C96.3506 29.3058 99.8526 32.7623 99.9857 37.023C99.9946 37.0807 99.9995 37.1412 100 37.2048V64.5355C100 64.607 99.9925 64.6772 99.9784 64.7455V90.7075C99.9674 95.1005 96.4618 98.7142 92.0474 98.7574C91.6153 98.7647 91.1878 98.7623 90.7589 98.76C90.5437 98.7586 90.3274 98.7574 90.1106 98.7574H7.94488C5.70526 98.7358 3.48723 97.7837 2.00491 96.0742C0.706642 94.5702 0.0249975 92.7094 0.0141764 90.7293V64.5345L0.014425 64.5087V49.0309C0.014425 47.7179 0.0108188 46.4051 0.00721254 45.0921C1.14721e-08 42.4666 -0.00721254 39.8411 0.014425 37.2156C0.046872 33.05 3.27119 29.5983 7.42599 29.2954C7.83981 29.2631 8.25365 29.267 8.66747 29.271C8.80543 29.2724 8.94346 29.2738 9.08142 29.2738H28.0782ZM48.906 19.0705C48.8936 18.0285 48.8809 16.9835 48.9152 15.9453C48.8155 14.1948 48.5283 12.4743 47.784 10.8583C46.1287 7.26614 42.9152 4.55042 39.0743 3.56579C35.2008 2.57024 30.9812 3.43581 27.8327 5.90278C24.6641 8.36848 22.8248 12.1853 22.8013 16.1863C22.8239 19.1494 23.8194 22.0906 25.7119 24.3938C27.3891 26.428 29.6178 27.9428 32.1497 28.6894C34.0542 29.256 36.0148 29.2405 37.9833 29.2248C38.3071 29.222 38.6314 29.2194 38.9551 29.2194H48.9189V20.6825C48.9189 20.147 48.9126 19.609 48.906 19.0705ZM61.924 29.23H51.0841V22.1434C51.0841 21.4745 51.0961 20.7988 51.1083 20.1197C51.131 18.8445 51.154 17.5571 51.0977 16.2809C51.1392 14.8561 51.2847 13.4716 51.7427 12.0806C52.9546 8.41273 55.9086 5.35065 59.5439 4.00903C63.3416 2.60245 67.6154 2.99198 71.0563 5.1343C74.8648 7.49303 77.1801 11.7125 77.2017 16.192C77.1691 21.4828 73.869 26.4599 68.8813 28.3534C67.6152 28.8295 66.2844 29.1326 64.9428 29.2082C64.2109 29.2476 63.4791 29.2412 62.7473 29.2347C62.473 29.2323 62.1985 29.23 61.924 29.23ZM95.5423 32.5955C96.992 33.6992 97.8036 35.4087 97.8252 37.194V63.4425H5.07803C4.65726 63.4425 4.23409 63.4367 3.80973 63.4306C3.26363 63.4231 2.71511 63.4151 2.16761 63.4203V37.2372C2.17842 35.8414 2.63285 34.4674 3.56335 33.4176C4.79682 32.0221 6.463 31.427 8.29149 31.427H91.539C92.978 31.427 94.3628 31.6976 95.5423 32.5955ZM97.8144 80.1569C97.2734 80.1588 96.7325 80.1557 96.1917 80.1527C95.7661 80.1503 95.3405 80.148 94.9149 80.148H2.16744V90.6319C2.16744 92.0601 2.56777 93.4559 3.51992 94.5486C4.64506 95.8471 6.22483 96.5829 7.94516 96.5936C9.53928 96.608 11.1334 96.6033 12.7275 96.5983C13.5246 96.596 14.3216 96.5936 15.1187 96.5936H92.0259C93.6488 96.5829 95.1636 95.9336 96.2778 94.7542C97.3273 93.6289 97.8144 92.1791 97.8144 90.6643V80.1569Z"
-      stroke-width="0.75"
-    />
-  </svg>
-</template>
