--- src/components/icons/scan-icon.vue
@@ -1,28 +1,0 @@
-<!--
-  - Copyright (c) TIKI Inc.
-  - MIT license. See LICENSE file in root directory.
-  -->
-
-<template>
-  <svg
-    viewBox="0 0 100 100"
-    fill="currentColor"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M3.57143 32.1429C4.51863 32.1429 5.42703 31.7666 6.09681 31.0968C6.76658 30.427 7.14286 29.5186 7.14286 28.5714V10.7143C7.14286 9.76708 7.51914 8.85868 8.1889 8.1889C8.85868 7.51914 9.76708 7.14286 10.7143 7.14286H30.3571C31.3043 7.14286 32.2127 6.76658 32.8825 6.09681C33.5523 5.42703 33.9286 4.51863 33.9286 3.57143C33.9286 2.62423 33.5523 1.71582 32.8825 1.04604C32.2127 0.376279 31.3043 0 30.3571 0H10.7143C7.87268 0 5.14746 1.12882 3.13815 3.13815C1.12882 5.14746 0 7.87268 0 10.7143V28.5714C0 29.5186 0.376279 30.427 1.04604 31.0968C1.71582 31.7666 2.62423 32.1429 3.57143 32.1429Z"
-    />
-    <path
-      d="M89.2857 0H69.6429C68.6957 0 67.7873 0.376279 67.1175 1.04604C66.4477 1.71582 66.0714 2.62423 66.0714 3.57143C66.0714 4.51863 66.4477 5.42703 67.1175 6.09681C67.7873 6.76658 68.6957 7.14286 69.6429 7.14286H89.2857C90.2329 7.14286 91.1413 7.51914 91.8111 8.1889C92.4809 8.85868 92.8571 9.76708 92.8571 10.7143V28.5714C92.8571 29.5186 93.2334 30.427 93.9032 31.0968C94.573 31.7666 95.4814 32.1429 96.4286 32.1429C97.3758 32.1429 98.2842 31.7666 98.954 31.0968C99.6237 30.427 100 29.5186 100 28.5714V10.7143C100 7.87268 98.8712 5.14746 96.8619 3.13815C94.8525 1.12882 92.1273 0 89.2857 0Z"
-    />
-    <path
-      d="M93.9032 68.9032C94.573 68.2334 95.4814 67.8571 96.4286 67.8571C97.3758 67.8571 98.2842 68.2334 98.954 68.9032C99.6237 69.573 100 70.4814 100 71.4286V89.2857C100 92.1273 98.8712 94.8525 96.8619 96.8619C94.8525 98.8712 92.1273 100 89.2857 100H69.6429C68.6957 100 67.7873 99.6237 67.1175 98.954C66.4477 98.2842 66.0714 97.3758 66.0714 96.4286C66.0714 95.4814 66.4477 94.573 67.1175 93.9032C67.7873 93.2334 68.6957 92.8571 69.6429 92.8571H89.2857C90.2329 92.8571 91.1413 92.4809 91.8111 91.8111C92.4809 91.1413 92.8571 90.2329 92.8571 89.2857V71.4286C92.8571 70.4814 93.2334 69.573 93.9032 68.9032Z"
-    />
-    <path
-      d="M30.3571 92.8571H10.7143C9.76708 92.8571 8.85868 92.4809 8.1889 91.8111C7.51914 91.1413 7.14286 90.2329 7.14286 89.2857V71.4286C7.14286 70.4814 6.76658 69.573 6.09681 68.9032C5.42703 68.2334 4.51863 67.8571 3.57143 67.8571C2.62423 67.8571 1.71582 68.2334 1.04604 68.9032C0.376279 69.573 0 70.4814 0 71.4286V89.2857C0 92.1273 1.12882 94.8525 3.13815 96.8619C5.14746 98.8712 7.87268 100 10.7143 100H30.3571C31.3043 100 32.2127 99.6237 32.8825 98.954C33.5523 98.2842 33.9286 97.3758 33.9286 96.4286C33.9286 95.4814 33.5523 94.573 32.8825 93.9032C32.2127 93.2334 31.3043 92.8571 30.3571 92.8571Z"
-    />
-    <path
-      d="M3.57143 46.4286H96.4286C97.3758 46.4286 98.2842 46.8048 98.954 47.4746C99.6237 48.1444 100 49.0528 100 50C100 50.9472 99.6237 51.8556 98.954 52.5254C98.2842 53.1952 97.3758 53.5714 96.4286 53.5714H3.57143C2.62423 53.5714 1.71582 53.1952 1.04604 52.5254C0.376279 51.8556 0 50.9472 0 50C0 49.0528 0.376279 48.1444 1.04604 47.4746C1.71582 46.8048 2.62423 46.4286 3.57143 46.4286Z"
-    />
-  </svg>
-</template>
