--- src/components/icons/unlink-icon.vue
@@ -1,34 +1,0 @@
-<!--
-  - Copyright (c) TIKI Inc.
-  - MIT license. See LICENSE file in root directory.
-  -->
-<script setup lang="ts">
-defineEmits(["click"]);
-</script>
-
-<template>
-  <svg
-    viewBox="0 0 100 100"
-    fill="'currentColor'"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M74.6975 0L66.362 16.2192L71.0629 18.6351L79.3983 2.41586L74.6975 0Z"
-    />
-    <path
-      d="M6.09503 32.4183C6.02128 39.3721 8.78934 46.0547 13.7593 50.9188L21.1592 58.3184L24.8601 54.6184L17.4594 47.218C12.1482 41.9068 10.0742 34.1658 12.0181 26.9105C13.9621 19.6553 19.6291 13.9882 26.8843 12.0443C34.1394 10.1003 41.8804 12.1743 47.1917 17.4855L54.5925 24.8863L58.2925 21.1863L50.8925 13.7856C46.0288 8.81553 39.3456 6.04736 32.392 6.12124C25.4068 6.08619 18.6975 8.84503 13.7587 13.7849C8.81884 18.7237 6.06013 25.433 6.09503 32.4183Z"
-    />
-    <path
-      d="M41.6415 78.801L45.3417 75.101L52.7421 82.5018C56.6642 86.4247 61.9938 88.6139 67.5417 88.5801C73.1075 88.5856 78.4471 86.3818 82.3896 82.4539C86.3317 78.5264 88.555 73.1939 88.5709 67.6293C88.5875 62.0639 86.3946 56.7193 82.4742 52.7697L75.0738 45.3689L78.7746 41.6689L86.1742 49.0684C91.065 54.0101 93.8009 60.6864 93.7825 67.6393C93.763 74.5909 90.9925 81.2535 86.0746 86.1685C81.158 91.0831 74.4938 93.8501 67.5417 93.866C60.5883 93.9397 53.9054 91.1714 49.0413 86.2014L41.6415 78.801Z"
-    />
-    <path
-      d="M97.5842 20.5485L81.365 28.8839L83.7808 33.585L100 25.2495L97.5842 20.5485Z"
-    />
-    <path
-      d="M28.9378 81.311L20.6024 97.5301L25.3034 99.946L33.6388 83.7268L28.9378 81.311Z"
-    />
-    <path
-      d="M0 74.648L16.2192 66.3126L18.6351 71.0138L2.41586 79.3492L0 74.648Z"
-    />
-  </svg>
-</template>
