+++ src/components/reward/reward-history.vue
@@ -4,11 +4,10 @@
   -->
 
 <script setup lang="ts">
-import ArrowIcon from "@/components/icons/arrow-icon.vue";
+import ArrowIcon from "@/assets/icons/arrow.svg";
 import { TikiService } from "@/service/tiki-service";
 import { inject, ref } from "vue";
 import type { HistoryEvent } from "@/service/history/history-event";
-import type { ReceiptEvent } from "@/service/receipt/receipt-event";
 
 defineEmits(["click"]);
 const tiki: TikiService | undefined = inject("Tiki");
